Move INotationConverter, CustomNotationConverter to /move_notation.
