+++ lib/shogi.dart
@@ -8,6 +8,6 @@ export 'src/models/board_piece.dart';
 export 'src/models/game_board.dart';
 export 'src/models/move.dart';
 export 'src/models/position.dart';
-export 'src/services/custom_notation_converter.dart';
 export 'src/services/game_engine.dart';
+export 'src/services/move_notation/custom_notation_converter.dart';
 export 'src/utils/shogi_utils.dart';
